pyproject.toml
[project]
name = "nefer-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi==0.110.0",
    "google-genai>=1.47.0",
    "jinja2==3.1.4",
    "markupsafe==2.1.5",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "supabase>=2.28.0",
    "trafilatura>=2.0.0",
    "uvicorn==0.27.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
